Validate that resource ids are valid mongodb object ids
This makes sure no mongoose errors are thrown if an invalid id is used.

+++ src/services/posts_service.js
@@ -1,6 +1,7 @@
 const Post = require("../models/post");
 const Comment = require("../models/comment");
 const AppError = require("../utils/app_error").AppError;
+const isValid = require("mongoose").Types.ObjectId.isValid;
 
 class PostsService {
    async getAllPosts() {
@@ -10,6 +11,10 @@ class PostsService {
    }
 
    async getPost(postId) {
+      if (!isValid(postId)) {
+         throw new AppError("Post not found", 404);
+      }
+
       const [post, comments] = await Promise.all([
          Post.find({ postId, isPublished: true }).populate("author", "name email isAuthor").exec(),
          Comment.find({ postId: postId }).populate("author", "name email isAuthor").exec(),
@@ -26,6 +31,10 @@ class PostsService {
    }
 
    async getPostComments(postId) {
+      if (!isValid(postId)) {
+         throw new AppError("Post not found", 404);
+      }
+
       return await Comment.find({ postId: postId })
          .populate("author", "name email isAuthor")
          .exec();
@@ -47,9 +56,7 @@ class PostsService {
    }
 
    async updatePost(currentUserId, authorId, title, text, isPublished, postId) {
-      const post = await Post.findById(postId).exec();
-
-      if (post === null) {
+      if (!isValid(postId)) {
          throw new AppError("Post not found", 404);
       }
 
@@ -63,28 +70,46 @@ class PostsService {
          isPublished,
       };
 
-      await Post.findByIdAndUpdate(postId, { $set: updatedPost }, { runValidators: true }).exec();
-   }
-
-   async deletePost(currentUserId, authorId, postId) {
-      const post = await Post.findById(postId).exec();
+      const post = await Post.findByIdAndUpdate(
+         postId,
+         { $set: updatedPost },
+         { runValidators: true }
+      ).exec();
 
       if (post === null) {
          throw new AppError("Post not found", 404);
       }
+   }
+
+   async deletePost(currentUserId, authorId, postId) {
+      if (!isValid(postId)) {
+         throw new AppError("Post not found", 404);
+      }
 
       if (currentUserId !== authorId) {
-         throw new AppError("You do not have the permission to delete this post", 403);
+         throw new AppError("You do not have the permission to update this post", 403);
       }
 
       const session = await Post.startSession();
-      session.startTransaction();
 
-      await Post.findByIdAndDelete(postId, { session }).exec();
-      await Comment.deleteMany({ postId }, { session }).exec();
+      try {
+         session.startTransaction();
 
-      await session.commitTransaction();
-      await session.endSession();
+         const post = await Post.findByIdAndDelete(postId, { session }).exec();
+
+         if (post === null) {
+            throw new AppError("Post not found", 404);
+         }
+
+         await Comment.deleteMany({ postId }, { session }).exec();
+
+         await session.commitTransaction();
+      } catch (error) {
+         await session.abortTransaction();
+         throw error;
+      } finally {
+         await session.endSession();
+      }
    }
 }
